[ML] Add clone tests for single metric, multi metric and population jobs (#46894)
This PR adds tests for ML single metric, multi metric and population job cloning.
