+++ x-pack/test/functional/apps/machine_learning/feature_controls/index.ts
@@ -7,7 +7,7 @@ import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default function({ loadTestFile }: FtrProviderContext) {
   describe('feature controls', function() {
-    this.tags('skipFirefox');
+    this.tags(['skipFirefox', 'mlqa']);
     loadTestFile(require.resolve('./ml_security'));
     loadTestFile(require.resolve('./ml_spaces'));
   });
