+++ x-pack/test/functional/apps/machine_learning/index.ts
@@ -11,8 +11,8 @@ export default function({ loadTestFile }: FtrProviderContext) {
 
     loadTestFile(require.resolve('./feature_controls'));
     loadTestFile(require.resolve('./pages'));
-    loadTestFile(require.resolve('./create_single_metric_job'));
-    loadTestFile(require.resolve('./create_multi_metric_job'));
-    loadTestFile(require.resolve('./create_population_job'));
+    loadTestFile(require.resolve('./single_metric_job'));
+    loadTestFile(require.resolve('./multi_metric_job'));
+    loadTestFile(require.resolve('./population_job'));
   });
 }
