+++ x-pack/test/functional/services/machine_learning/job_wizard_common.ts
@@ -13,6 +13,18 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
   const testSubjects = getService('testSubjects');
 
   return {
+    async waitForNextButtonVisible() {
+      await retry.waitFor(
+        'next button to be visible',
+        async () => await testSubjects.isDisplayed('mlJobWizardNavButtonNext')
+      );
+    },
+
+    async clickNextButton() {
+      await this.waitForNextButtonVisible();
+      await testSubjects.clickWhenNotDisabled('mlJobWizardNavButtonNext');
+    },
+
     async assertTimeRangeSectionExists() {
       await testSubjects.existOrFail('mlJobWizardStepTitleTimeRange');
     },
@@ -33,20 +45,49 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
       await testSubjects.existOrFail('mlJobWizardStepTitleSummary');
     },
 
+    async advanceToPickFieldsSection() {
+      await this.clickNextButton();
+      await this.assertPickFieldsSectionExists();
+    },
+
+    async advanceToJobDetailsSection() {
+      await this.clickNextButton();
+      await this.assertJobDetailsSectionExists();
+    },
+
+    async advanceToValidationSection() {
+      await this.clickNextButton();
+      await this.assertValidationSectionExists();
+    },
+
+    async advanceToSummarySection() {
+      await this.clickNextButton();
+      await this.assertSummarySectionExists();
+    },
+
     async assertEventRateChartExists() {
-      await testSubjects.existOrFail('mlEventRateChart');
+      await testSubjects.existOrFail('~mlEventRateChart');
+    },
+
+    async assertEventRateChartHasData() {
+      await testSubjects.existOrFail('mlEventRateChart withData');
     },
 
     async assertAggAndFieldInputExists() {
       await testSubjects.existOrFail('mlJobWizardAggSelection > comboBoxInput');
     },
 
-    async assertAggAndFieldSelection(identifier: string) {
+    async assertAggAndFieldSelection(expectedIdentifier: string) {
       const comboBoxSelectedOptions = await comboBox.getComboBoxSelectedOptions(
         'mlJobWizardAggSelection > comboBoxInput'
       );
       expect(comboBoxSelectedOptions.length).to.eql(1);
-      expect(comboBoxSelectedOptions[0]).to.eql(identifier);
+      expect(comboBoxSelectedOptions[0]).to.eql(expectedIdentifier);
+    },
+
+    async selectAggAndField(identifier: string, isIdentifierKeptInField: boolean) {
+      await comboBox.set('mlJobWizardAggSelection > comboBoxInput', identifier);
+      await this.assertAggAndFieldSelection(isIdentifierKeptInField ? identifier : '');
     },
 
     async assertBucketSpanInputExists() {
@@ -61,6 +102,11 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
       expect(actualBucketSpan).to.eql(expectedValue);
     },
 
+    async setBucketSpan(bucketSpan: string) {
+      await testSubjects.setValue('mlJobWizardInputBucketSpan', bucketSpan, { withKeyboard: true });
+      await this.assertBucketSpanValue(bucketSpan);
+    },
+
     async assertJobIdInputExists() {
       await testSubjects.existOrFail('mlJobWizardInputJobId');
     },
@@ -68,6 +114,11 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
     async assertJobIdValue(expectedValue: string) {
       const actualJobId = await testSubjects.getAttribute('mlJobWizardInputJobId', 'value');
       expect(actualJobId).to.eql(expectedValue);
+    },
+
+    async setJobId(jobId: string) {
+      await testSubjects.setValue('mlJobWizardInputJobId', jobId, { withKeyboard: true });
+      await this.assertJobIdValue(jobId);
     },
 
     async assertJobDescriptionInputExists() {
@@ -81,15 +132,30 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
       expect(actualJobDescription).to.eql(expectedValue);
     },
 
+    async setJobDescription(jobDescription: string) {
+      await testSubjects.setValue('mlJobWizardInputJobDescription', jobDescription, {
+        withKeyboard: true,
+      });
+      await this.assertJobDescriptionValue(jobDescription);
+    },
+
     async assertJobGroupInputExists() {
       await testSubjects.existOrFail('mlJobWizardComboBoxJobGroups > comboBoxInput');
     },
 
-    async assertJobGroupSelection(jobGroups: string[]) {
-      const comboBoxSelectedOptions = await comboBox.getComboBoxSelectedOptions(
+    async getSelectedJobGroups(): Promise<string[]> {
+      return await comboBox.getComboBoxSelectedOptions(
         'mlJobWizardComboBoxJobGroups > comboBoxInput'
       );
-      expect(comboBoxSelectedOptions).to.eql(jobGroups);
+    },
+
+    async assertJobGroupSelection(jobGroups: string[]) {
+      expect(await this.getSelectedJobGroups()).to.eql(jobGroups);
+    },
+
+    async addJobGroup(jobGroup: string) {
+      await comboBox.setCustom('mlJobWizardComboBoxJobGroups > comboBoxInput', jobGroup);
+      expect(await this.getSelectedJobGroups()).to.contain(jobGroup);
     },
 
     async assertModelPlotSwitchExists() {
@@ -97,6 +163,19 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
       await testSubjects.existOrFail('mlJobWizardAdvancedSection > mlJobWizardSwitchModelPlot', {
         allowHidden: true,
       });
+    },
+
+    async getModelPlotSwitchCheckedState(): Promise<boolean> {
+      await this.ensureAdvancedSectionOpen();
+      return await testSubjects.isSelected(
+        'mlJobWizardAdvancedSection > mlJobWizardSwitchModelPlot'
+      );
+    },
+
+    async assertModelPlotSwitchCheckedState(expectedValue: boolean) {
+      await this.ensureAdvancedSectionOpen();
+      const actualCheckedState = await this.getModelPlotSwitchCheckedState();
+      expect(actualCheckedState).to.eql(expectedValue);
     },
 
     async assertDedicatedIndexSwitchExists() {
@@ -107,21 +186,24 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
       );
     },
 
+    async getDedicatedIndexSwitchCheckedState(): Promise<boolean> {
+      await this.ensureAdvancedSectionOpen();
+      return await testSubjects.isSelected(
+        'mlJobWizardAdvancedSection > mlJobWizardSwitchUseDedicatedIndex'
+      );
+    },
+
     async assertDedicatedIndexSwitchCheckedState(expectedValue: boolean) {
       await this.ensureAdvancedSectionOpen();
       const actualCheckedState = await this.getDedicatedIndexSwitchCheckedState();
       expect(actualCheckedState).to.eql(expectedValue);
     },
 
-    async assertCreateJobButtonExists() {
-      await testSubjects.existOrFail('mlJobWizardButtonCreateJob');
-    },
-
-    async getDedicatedIndexSwitchCheckedState() {
-      await this.ensureAdvancedSectionOpen();
-      return await testSubjects.isSelected(
-        'mlJobWizardAdvancedSection > mlJobWizardSwitchUseDedicatedIndex'
-      );
+    async activateDedicatedIndexSwitch() {
+      if ((await this.getDedicatedIndexSwitchCheckedState()) === false) {
+        await testSubjects.clickWhenNotDisabled('mlJobWizardSwitchUseDedicatedIndex');
+      }
+      await this.assertDedicatedIndexSwitchCheckedState(true);
     },
 
     async assertModelMemoryLimitInputExists() {
@@ -140,15 +222,33 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
       expect(actualModelMemoryLimit).to.eql(expectedValue);
     },
 
+    async setModelMemoryLimit(modelMemoryLimit: string) {
+      await testSubjects.setValue('mlJobWizardInputModelMemoryLimit', modelMemoryLimit, {
+        withKeyboard: true,
+      });
+      await this.assertModelMemoryLimitValue(modelMemoryLimit);
+    },
+
     async assertInfluencerInputExists() {
       await testSubjects.existOrFail('influencerSelect > comboBoxInput');
     },
 
+    async getSelectedInfluencers(): Promise<string[]> {
+      return await comboBox.getComboBoxSelectedOptions('influencerSelect > comboBoxInput');
+    },
+
     async assertInfluencerSelection(influencers: string[]) {
-      const comboBoxSelectedOptions = await comboBox.getComboBoxSelectedOptions(
-        'influencerSelect > comboBoxInput'
-      );
-      expect(comboBoxSelectedOptions).to.eql(influencers);
+      expect(await this.getSelectedInfluencers()).to.eql(influencers);
+    },
+
+    async addInfluencer(influencer: string) {
+      await comboBox.setCustom('influencerSelect > comboBoxInput', influencer);
+      expect(await this.getSelectedInfluencers()).to.contain(influencer);
+    },
+
+    async assertAnomalyChartExists(chartType: string, preSelector?: string) {
+      let chartSelector = `mlAnomalyChart ${chartType}`;
+      chartSelector = !preSelector ? chartSelector : `${preSelector} > ${chartSelector}`;
     },
 
     async assertDetectorPreviewExists(
@@ -161,45 +261,55 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
       expect(
         await testSubjects.getVisibleText(`detector ${detectorPosition} > detectorTitle`)
       ).to.eql(aggAndFieldIdentifier);
-      await testSubjects.existOrFail(`detector ${detectorPosition} > mlAnomalyChart ${chartType}`);
+
+      await this.assertAnomalyChartExists(chartType, `detector ${detectorPosition}`);
     },
 
-    async clickNextButton() {
-      await testSubjects.clickWhenNotDisabled('mlJobWizardNavButtonNext');
+    async assertDetectorSplitExists(splitField: string) {
+      await testSubjects.existOrFail(`dataSplit > dataSplitTitle ${splitField}`);
+      await testSubjects.existOrFail(`dataSplit > splitCard front`);
+      await testSubjects.existOrFail(`dataSplit > splitCard back`);
     },
 
-    async clickPreviousButton() {
-      await testSubjects.clickWhenNotDisabled('mlJobWizardNavButtonPrevious');
+    async assertDetectorSplitFrontCardTitle(frontCardTitle: string) {
+      expect(
+        await testSubjects.getVisibleText(`dataSplit > splitCard front > splitCardTitle`)
+      ).to.eql(frontCardTitle);
     },
 
-    async clickUseFullDataButton() {
-      await testSubjects.clickWhenNotDisabled('mlButtonUseFullData');
+    async assertDetectorSplitNumberOfBackCards(numberOfBackCards: number) {
+      expect(await testSubjects.findAll(`dataSplit > splitCard back`)).to.have.length(
+        numberOfBackCards
+      );
     },
 
-    async selectAggAndField(identifier: string) {
-      await comboBox.set('mlJobWizardAggSelection > comboBoxInput', identifier);
+    async assertCreateJobButtonExists() {
+      await testSubjects.existOrFail('mlJobWizardButtonCreateJob');
     },
 
-    async setBucketSpan(bucketSpan: string) {
-      await testSubjects.setValue('mlJobWizardInputBucketSpan', bucketSpan, { withKeyboard: true });
+    async assertDateRangeSelectionExists() {
+      await testSubjects.existOrFail('jobWizardDateRange');
     },
 
-    async setJobId(jobId: string) {
-      await testSubjects.setValue('mlJobWizardInputJobId', jobId, { withKeyboard: true });
+    async getSelectedDateRange() {
+      const dateRange = await testSubjects.find('jobWizardDateRange');
+      const [startPicker, endPicker] = await dateRange.findAllByClassName('euiFieldText');
+      return {
+        startDate: await startPicker.getAttribute('value'),
+        endDate: await endPicker.getAttribute('value'),
+      };
     },
 
-    async setJobDescription(jobDescription: string) {
-      await testSubjects.setValue('mlJobWizardInputJobDescription', jobDescription, {
-        withKeyboard: true,
+    async assertDateRangeSelection(expectedStartDate: string, expectedEndDate: string) {
+      expect(await this.getSelectedDateRange()).to.eql({
+        startDate: expectedStartDate,
+        endDate: expectedEndDate,
       });
     },
 
-    async addJobGroup(jobGroup: string) {
-      await comboBox.setCustom('mlJobWizardComboBoxJobGroups > comboBoxInput', jobGroup);
-    },
-
-    async addInfluencer(influencer: string) {
-      await comboBox.setCustom('influencerSelect > comboBoxInput', influencer);
+    async clickUseFullDataButton(expectedStartDate: string, expectedEndDate: string) {
+      await testSubjects.clickWhenNotDisabled('mlButtonUseFullData');
+      await this.assertDateRangeSelection(expectedStartDate, expectedEndDate);
     },
 
     async ensureAdvancedSectionOpen() {
@@ -211,24 +321,12 @@ export function MachineLearningJobWizardCommonProvider({ getService }: FtrProvid
       });
     },
 
-    async activateDedicatedIndexSwitch() {
-      if ((await this.getDedicatedIndexSwitchCheckedState()) === false) {
-        await testSubjects.clickWhenNotDisabled('mlJobWizardSwitchUseDedicatedIndex');
-      }
-    },
-
-    async setModelMemoryLimit(modelMemoryLimit: string) {
-      await testSubjects.setValue('mlJobWizardInputModelMemoryLimit', modelMemoryLimit, {
-        withKeyboard: true,
-      });
-    },
-
     async createJobAndWaitForCompletion() {
       await testSubjects.clickWhenNotDisabled('mlJobWizardButtonCreateJob');
       await retry.waitForWithTimeout(
         'job processing to finish',
         5 * 60 * 1000,
-        async () => (await testSubjects.exists('mlJobWizardButtonCreateJob')) === false
+        async () => await testSubjects.exists('mlJobWizardButtonRunInRealTime')
       );
     },
   };
